@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp2.2</TargetFramework>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
         <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
         <TypeScriptToolsVersion>Latest</TypeScriptToolsVersion>
         <IsPackable>false</IsPackable>
@@ -21,15 +21,24 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="FluentValidation.AspNetCore" Version="8.4.0" />
-        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.8.2" />
-        <PackageReference Include="Microsoft.AspNetCore.App" />
-        <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
-        <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="2.2.0" />
-        <PackageReference Include="Polly" Version="7.1.0" />
-        <PackageReference Include="SonarAnalyzer.CSharp" Version="7.14.0.8411" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
-        <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="4.0.1" />
+        <PackageReference Include="FluentValidation.AspNetCore" Version="8.6.1" />
+        <PackageReference Include="MicroElements.Swashbuckle.FluentValidation" Version="3.0.0-rc.4" />
+        <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.12.0" />
+        <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.Http.Connections.Common" Version="3.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.SignalR.Protocols.NewtonsoftJson" Version="3.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.SpaServices" Version="3.1.1" />
+        <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="3.1.1" />
+        <PackageReference Include="Microsoft.Extensions.DependencyModel" Version="3.1.1" />
+        <PackageReference Include="Microsoft.Extensions.Logging.ApplicationInsights" Version="2.12.0" />
+        <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="3.1.1" />
+        <PackageReference Include="Microsoft.OpenApi" Version="1.1.4" />
+        <PackageReference Include="Polly" Version="7.2.0" />
+        <PackageReference Include="SonarAnalyzer.CSharp" Version="8.2.0.14119" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0" />
+        <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="5.0.0" />
+        <PackageReference Include="Swashbuckle.AspNetCore.Newtonsoft" Version="5.0.0" />
     </ItemGroup>
 
     <ItemGroup>
