@@ -1,6 +1,6 @@
 ﻿//----------------------
 // <auto-generated>
-//     Generated using the NSwag toolchain v13.0.4.0 (NJsonSchema v10.0.21.0 (Newtonsoft.Json v11.0.0.0)) (http://NSwag.org)
+//     Generated using the NSwag toolchain v13.2.2.0 (NJsonSchema v10.1.4.0 (Newtonsoft.Json v12.0.0.0)) (http://NSwag.org)
 // </auto-generated>
 //----------------------
 
@@ -12,7 +12,9 @@
 
 namespace VideoWeb.Services.User
 {
-    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.0.4.0 (NJsonSchema v10.0.21.0 (Newtonsoft.Json v11.0.0.0))")]
+    using System = global::System;
+    
+    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.2.2.0 (NJsonSchema v10.1.4.0 (Newtonsoft.Json v12.0.0.0))")]
     public partial interface IUserApiClient
     {
         /// <summary>Get AD Group By Name</summary>
@@ -66,18 +68,18 @@ namespace VideoWeb.Services.User
         /// <summary>Add a user to a group</summary>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest request);
+        System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest body);
     
         /// <summary>Add a user to a group</summary>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        void AddUserToGroup(AddUserToGroupRequest request);
+        void AddUserToGroup(AddUserToGroupRequest body);
     
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <summary>Add a user to a group</summary>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest request, System.Threading.CancellationToken cancellationToken);
+        System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest body, System.Threading.CancellationToken cancellationToken);
     
         /// <summary>Run a health check of the service</summary>
         /// <returns>Success</returns>
@@ -96,23 +98,39 @@ namespace VideoWeb.Services.User
         System.Threading.Tasks.Task CheckServiceHealthAsync(System.Threading.CancellationToken cancellationToken);
     
         /// <summary>Create a new hearings reforms user account</summary>
-        /// <param name="request">Details of a new user</param>
+        /// <param name="body">Details of a new user</param>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest request);
+        System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest body);
     
         /// <summary>Create a new hearings reforms user account</summary>
-        /// <param name="request">Details of a new user</param>
+        /// <param name="body">Details of a new user</param>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        NewUserResponse CreateUser(CreateUserRequest request);
+        NewUserResponse CreateUser(CreateUserRequest body);
     
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <summary>Create a new hearings reforms user account</summary>
-        /// <param name="request">Details of a new user</param>
+        /// <param name="body">Details of a new user</param>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest request, System.Threading.CancellationToken cancellationToken);
+        System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest body, System.Threading.CancellationToken cancellationToken);
+    
+        /// <summary>Updates an AAD user</summary>
+        /// <returns>Success</returns>
+        /// <exception cref="UserApiException">A server side error occurred.</exception>
+        System.Threading.Tasks.Task UpdateUserAsync(string body);
+    
+        /// <summary>Updates an AAD user</summary>
+        /// <returns>Success</returns>
+        /// <exception cref="UserApiException">A server side error occurred.</exception>
+        void UpdateUser(string body);
+    
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <summary>Updates an AAD user</summary>
+        /// <returns>Success</returns>
+        /// <exception cref="UserApiException">A server side error occurred.</exception>
+        System.Threading.Tasks.Task UpdateUserAsync(string body, System.Threading.CancellationToken cancellationToken);
     
         /// <summary>Get User by AD User ID</summary>
         /// <returns>Success</returns>
@@ -196,7 +214,7 @@ namespace VideoWeb.Services.User
     
     }
     
-    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.0.4.0 (NJsonSchema v10.0.21.0 (Newtonsoft.Json v11.0.0.0))")]
+    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.2.2.0 (NJsonSchema v10.1.4.0 (Newtonsoft.Json v12.0.0.0))")]
     public partial class UserApiClient : IUserApiClient
     {
         private string _baseUrl = "http://localhost:5400";
@@ -206,12 +224,14 @@ namespace VideoWeb.Services.User
         public UserApiClient(System.Net.Http.HttpClient httpClient)
         {
             _httpClient = httpClient; 
-            _settings = new System.Lazy<Newtonsoft.Json.JsonSerializerSettings>(() => 
-            {
-                var settings = new Newtonsoft.Json.JsonSerializerSettings();
-                UpdateJsonSerializerSettings(settings);
-                return settings;
-            });
+            _settings = new System.Lazy<Newtonsoft.Json.JsonSerializerSettings>(CreateSerializerSettings);
+        }
+    
+        private Newtonsoft.Json.JsonSerializerSettings CreateSerializerSettings()
+        {
+            var settings = new Newtonsoft.Json.JsonSerializerSettings();
+            UpdateJsonSerializerSettings(settings);
+            return settings;
         }
     
         public string BaseUrl 
@@ -219,6 +239,7 @@ namespace VideoWeb.Services.User
             get { return _baseUrl; }
             set { _baseUrl = value; }
         }
+    
     
         public Newtonsoft.Json.JsonSerializerSettings JsonSerializerSettings { get { return _settings.Value; } }
     
@@ -253,7 +274,7 @@ namespace VideoWeb.Services.User
             urlBuilder_.Append(BaseUrl != null ? BaseUrl.TrimEnd('/') : "").Append("/accounts/group?");
             if (name != null) 
             {
-                urlBuilder_.Append("name=").Append(System.Uri.EscapeDataString(ConvertToString(name, System.Globalization.CultureInfo.InvariantCulture))).Append("&");
+                urlBuilder_.Append(System.Uri.EscapeDataString("name") + "=").Append(System.Uri.EscapeDataString(ConvertToString(name, System.Globalization.CultureInfo.InvariantCulture))).Append("&");
             }
             urlBuilder_.Length--;
     
@@ -291,13 +312,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -384,13 +405,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -477,13 +498,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -510,24 +531,24 @@ namespace VideoWeb.Services.User
         /// <summary>Add a user to a group</summary>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        public System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest request)
+        public System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest body)
         {
-            return AddUserToGroupAsync(request, System.Threading.CancellationToken.None);
+            return AddUserToGroupAsync(body, System.Threading.CancellationToken.None);
         }
     
         /// <summary>Add a user to a group</summary>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        public void AddUserToGroup(AddUserToGroupRequest request)
+        public void AddUserToGroup(AddUserToGroupRequest body)
         {
-            System.Threading.Tasks.Task.Run(async () => await AddUserToGroupAsync(request, System.Threading.CancellationToken.None)).GetAwaiter().GetResult();
+            System.Threading.Tasks.Task.Run(async () => await AddUserToGroupAsync(body, System.Threading.CancellationToken.None)).GetAwaiter().GetResult();
         }
     
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <summary>Add a user to a group</summary>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        public async System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest request, System.Threading.CancellationToken cancellationToken)
+        public async System.Threading.Tasks.Task AddUserToGroupAsync(AddUserToGroupRequest body, System.Threading.CancellationToken cancellationToken)
         {
             var urlBuilder_ = new System.Text.StringBuilder();
             urlBuilder_.Append(BaseUrl != null ? BaseUrl.TrimEnd('/') : "").Append("/accounts/user/group");
@@ -537,8 +558,8 @@ namespace VideoWeb.Services.User
             {
                 using (var request_ = new System.Net.Http.HttpRequestMessage())
                 {
-                    var content_ = new System.Net.Http.StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(request, _settings.Value));
-                    content_.Headers.ContentType = System.Net.Http.Headers.MediaTypeHeaderValue.Parse("application/json");
+                    var content_ = new System.Net.Http.StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(body, _settings.Value));
+                    content_.Headers.ContentType = System.Net.Http.Headers.MediaTypeHeaderValue.Parse("application/json-patch+json");
                     request_.Content = content_;
                     request_.Method = new System.Net.Http.HttpMethod("PATCH");
     
@@ -567,19 +588,19 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "400") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Bad Request", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Bad Request", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -658,8 +679,14 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "500") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Server Error", (int)response_.StatusCode, responseText_, headers_, null);
+                        }
+                        else
+                        if (status_ == "401") 
+                        {
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
                         if (status_ != "200" && status_ != "204")
@@ -681,29 +708,29 @@ namespace VideoWeb.Services.User
         }
     
         /// <summary>Create a new hearings reforms user account</summary>
-        /// <param name="request">Details of a new user</param>
+        /// <param name="body">Details of a new user</param>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        public System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest request)
+        public System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest body)
         {
-            return CreateUserAsync(request, System.Threading.CancellationToken.None);
+            return CreateUserAsync(body, System.Threading.CancellationToken.None);
         }
     
         /// <summary>Create a new hearings reforms user account</summary>
-        /// <param name="request">Details of a new user</param>
+        /// <param name="body">Details of a new user</param>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        public NewUserResponse CreateUser(CreateUserRequest request)
+        public NewUserResponse CreateUser(CreateUserRequest body)
         {
-            return System.Threading.Tasks.Task.Run(async () => await CreateUserAsync(request, System.Threading.CancellationToken.None)).GetAwaiter().GetResult();
+            return System.Threading.Tasks.Task.Run(async () => await CreateUserAsync(body, System.Threading.CancellationToken.None)).GetAwaiter().GetResult();
         }
     
         /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
         /// <summary>Create a new hearings reforms user account</summary>
-        /// <param name="request">Details of a new user</param>
+        /// <param name="body">Details of a new user</param>
         /// <returns>Success</returns>
         /// <exception cref="UserApiException">A server side error occurred.</exception>
-        public async System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest request, System.Threading.CancellationToken cancellationToken)
+        public async System.Threading.Tasks.Task<NewUserResponse> CreateUserAsync(CreateUserRequest body, System.Threading.CancellationToken cancellationToken)
         {
             var urlBuilder_ = new System.Text.StringBuilder();
             urlBuilder_.Append(BaseUrl != null ? BaseUrl.TrimEnd('/') : "").Append("/users");
@@ -713,8 +740,8 @@ namespace VideoWeb.Services.User
             {
                 using (var request_ = new System.Net.Http.HttpRequestMessage())
                 {
-                    var content_ = new System.Net.Http.StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(request, _settings.Value));
-                    content_.Headers.ContentType = System.Net.Http.Headers.MediaTypeHeaderValue.Parse("application/json");
+                    var content_ = new System.Net.Http.StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(body, _settings.Value));
+                    content_.Headers.ContentType = System.Net.Http.Headers.MediaTypeHeaderValue.Parse("application/json-patch+json");
                     request_.Content = content_;
                     request_.Method = new System.Net.Http.HttpMethod("POST");
                     request_.Headers.Accept.Add(System.Net.Http.Headers.MediaTypeWithQualityHeaderValue.Parse("application/json"));
@@ -745,13 +772,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -762,6 +789,100 @@ namespace VideoWeb.Services.User
                         }
             
                         return default(NewUserResponse);
+                    }
+                    finally
+                    {
+                        if (response_ != null)
+                            response_.Dispose();
+                    }
+                }
+            }
+            finally
+            {
+            }
+        }
+    
+        /// <summary>Updates an AAD user</summary>
+        /// <returns>Success</returns>
+        /// <exception cref="UserApiException">A server side error occurred.</exception>
+        public System.Threading.Tasks.Task UpdateUserAsync(string body)
+        {
+            return UpdateUserAsync(body, System.Threading.CancellationToken.None);
+        }
+    
+        /// <summary>Updates an AAD user</summary>
+        /// <returns>Success</returns>
+        /// <exception cref="UserApiException">A server side error occurred.</exception>
+        public void UpdateUser(string body)
+        {
+            System.Threading.Tasks.Task.Run(async () => await UpdateUserAsync(body, System.Threading.CancellationToken.None)).GetAwaiter().GetResult();
+        }
+    
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <summary>Updates an AAD user</summary>
+        /// <returns>Success</returns>
+        /// <exception cref="UserApiException">A server side error occurred.</exception>
+        public async System.Threading.Tasks.Task UpdateUserAsync(string body, System.Threading.CancellationToken cancellationToken)
+        {
+            var urlBuilder_ = new System.Text.StringBuilder();
+            urlBuilder_.Append(BaseUrl != null ? BaseUrl.TrimEnd('/') : "").Append("/users");
+    
+            var client_ = _httpClient;
+            try
+            {
+                using (var request_ = new System.Net.Http.HttpRequestMessage())
+                {
+                    var content_ = new System.Net.Http.StringContent(Newtonsoft.Json.JsonConvert.SerializeObject(body, _settings.Value));
+                    content_.Headers.ContentType = System.Net.Http.Headers.MediaTypeHeaderValue.Parse("application/json-patch+json");
+                    request_.Content = content_;
+                    request_.Method = new System.Net.Http.HttpMethod("PATCH");
+    
+                    PrepareRequest(client_, request_, urlBuilder_);
+                    var url_ = urlBuilder_.ToString();
+                    request_.RequestUri = new System.Uri(url_, System.UriKind.RelativeOrAbsolute);
+                    PrepareRequest(client_, request_, url_);
+    
+                    var response_ = await client_.SendAsync(request_, System.Net.Http.HttpCompletionOption.ResponseHeadersRead, cancellationToken).ConfigureAwait(false);
+                    try
+                    {
+                        var headers_ = System.Linq.Enumerable.ToDictionary(response_.Headers, h_ => h_.Key, h_ => h_.Value);
+                        if (response_.Content != null && response_.Content.Headers != null)
+                        {
+                            foreach (var item_ in response_.Content.Headers)
+                                headers_[item_.Key] = item_.Value;
+                        }
+    
+                        ProcessResponse(client_, response_);
+    
+                        var status_ = ((int)response_.StatusCode).ToString();
+                        if (status_ == "204") 
+                        {
+                            return;
+                        }
+                        else
+                        if (status_ == "400") 
+                        {
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Bad Request", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
+                        }
+                        else
+                        if (status_ == "404") 
+                        {
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
+                        }
+                        else
+                        if (status_ == "401") 
+                        {
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
+                        }
+                        else
+                        if (status_ != "200" && status_ != "204")
+                        {
+                            var responseData_ = response_.Content == null ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false); 
+                            throw new UserApiException("The HTTP status code of the response was not expected (" + (int)response_.StatusCode + ").", (int)response_.StatusCode, responseData_, headers_, null);
+                        }
                     }
                     finally
                     {
@@ -838,13 +959,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -931,13 +1052,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -1022,19 +1143,19 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "400") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Bad Request", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Bad Request", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -1119,13 +1240,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -1208,13 +1329,13 @@ namespace VideoWeb.Services.User
                         else
                         if (status_ == "404") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
-                            throw new UserApiException("Not Found", (int)response_.StatusCode, responseText_, headers_, null);
+                            var objectResponse_ = await ReadObjectResponseAsync<ProblemDetails>(response_, headers_).ConfigureAwait(false);
+                            throw new UserApiException<ProblemDetails>("Not Found", (int)response_.StatusCode, objectResponse_.Text, headers_, objectResponse_.Object, null);
                         }
                         else
                         if (status_ == "401") 
                         {
-                            string responseText_ = ( response_.Content == null ) ? null : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
+                            string responseText_ = ( response_.Content == null ) ? string.Empty : await response_.Content.ReadAsStringAsync().ConfigureAwait(false);
                             throw new UserApiException("Unauthorized", (int)response_.StatusCode, responseText_, headers_, null);
                         }
                         else
@@ -1331,7 +1452,7 @@ namespace VideoWeb.Services.User
         }
     }
 
-    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.0.21.0 (Newtonsoft.Json v11.0.0.0)")]
+    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.1.4.0 (Newtonsoft.Json v12.0.0.0)")]
     public partial class GroupsResponse 
     {
         [Newtonsoft.Json.JsonProperty("group_id", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
@@ -1343,34 +1464,58 @@ namespace VideoWeb.Services.User
     
     }
     
-    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.0.21.0 (Newtonsoft.Json v11.0.0.0)")]
+    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.1.4.0 (Newtonsoft.Json v12.0.0.0)")]
+    public partial class ProblemDetails 
+    {
+        [Newtonsoft.Json.JsonProperty("type", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        public string Type { get; set; }
+    
+        [Newtonsoft.Json.JsonProperty("title", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        public string Title { get; set; }
+    
+        [Newtonsoft.Json.JsonProperty("status", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        public int? Status { get; set; }
+    
+        [Newtonsoft.Json.JsonProperty("detail", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        public string Detail { get; set; }
+    
+        [Newtonsoft.Json.JsonProperty("instance", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        public string Instance { get; set; }
+    
+        [Newtonsoft.Json.JsonProperty("extensions", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        public System.Collections.Generic.Dictionary<string, object> Extensions { get; set; }
+    
+    
+    }
+    
+    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.1.4.0 (Newtonsoft.Json v12.0.0.0)")]
     public partial class AddUserToGroupRequest 
     {
-        [Newtonsoft.Json.JsonProperty("user_id", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        [Newtonsoft.Json.JsonProperty("user_id", Required = Newtonsoft.Json.Required.AllowNull)]
         public string User_id { get; set; }
     
-        [Newtonsoft.Json.JsonProperty("group_name", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        [Newtonsoft.Json.JsonProperty("group_name", Required = Newtonsoft.Json.Required.AllowNull)]
         public string Group_name { get; set; }
     
     
     }
     
-    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.0.21.0 (Newtonsoft.Json v11.0.0.0)")]
+    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.1.4.0 (Newtonsoft.Json v12.0.0.0)")]
     public partial class CreateUserRequest 
     {
-        [Newtonsoft.Json.JsonProperty("first_name", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        [Newtonsoft.Json.JsonProperty("first_name", Required = Newtonsoft.Json.Required.AllowNull)]
         public string First_name { get; set; }
     
-        [Newtonsoft.Json.JsonProperty("last_name", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        [Newtonsoft.Json.JsonProperty("last_name", Required = Newtonsoft.Json.Required.AllowNull)]
         public string Last_name { get; set; }
     
-        [Newtonsoft.Json.JsonProperty("recovery_email", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
+        [Newtonsoft.Json.JsonProperty("recovery_email", Required = Newtonsoft.Json.Required.AllowNull)]
         public string Recovery_email { get; set; }
     
     
     }
     
-    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.0.21.0 (Newtonsoft.Json v11.0.0.0)")]
+    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.1.4.0 (Newtonsoft.Json v12.0.0.0)")]
     public partial class NewUserResponse 
     {
         [Newtonsoft.Json.JsonProperty("user_id", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
@@ -1385,7 +1530,7 @@ namespace VideoWeb.Services.User
     
     }
     
-    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.0.21.0 (Newtonsoft.Json v11.0.0.0)")]
+    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.1.4.0 (Newtonsoft.Json v12.0.0.0)")]
     public partial class UserProfile 
     {
         [Newtonsoft.Json.JsonProperty("user_id", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
@@ -1416,18 +1561,15 @@ namespace VideoWeb.Services.User
     }
     
     /// <summary>A judge existing in the system</summary>
-    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.0.21.0 (Newtonsoft.Json v11.0.0.0)")]
+    [System.CodeDom.Compiler.GeneratedCode("NJsonSchema", "10.1.4.0 (Newtonsoft.Json v12.0.0.0)")]
     public partial class UserResponse 
     {
-        /// <summary>Judge first name</summary>
         [Newtonsoft.Json.JsonProperty("first_name", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string First_name { get; set; }
     
-        /// <summary>Judge last name</summary>
         [Newtonsoft.Json.JsonProperty("last_name", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string Last_name { get; set; }
     
-        /// <summary>Judge display name as in the identity system</summary>
         [Newtonsoft.Json.JsonProperty("display_name", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string Display_name { get; set; }
     
@@ -1438,7 +1580,7 @@ namespace VideoWeb.Services.User
     
     }
 
-    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.0.4.0 (NJsonSchema v10.0.21.0 (Newtonsoft.Json v11.0.0.0))")]
+    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.2.2.0 (NJsonSchema v10.1.4.0 (Newtonsoft.Json v12.0.0.0))")]
     public partial class UserApiException : System.Exception
     {
         public int StatusCode { get; private set; }
@@ -1461,7 +1603,7 @@ namespace VideoWeb.Services.User
         }
     }
 
-    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.0.4.0 (NJsonSchema v10.0.21.0 (Newtonsoft.Json v11.0.0.0))")]
+    [System.CodeDom.Compiler.GeneratedCode("NSwag", "13.2.2.0 (NJsonSchema v10.1.4.0 (Newtonsoft.Json v12.0.0.0))")]
     public partial class UserApiException<TResult> : UserApiException
     {
         public TResult Result { get; private set; }
