@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace VideoWeb.Contract.Responses
+﻿namespace VideoWeb.Contract.Responses
 {
     public class SelfTestPexipResponse
     {
