@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.SignalR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyModel;
 using Microsoft.Extensions.Options;
+using Microsoft.OpenApi.Models;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Serialization;
@@ -19,6 +20,7 @@ using VideoWeb.Common;
 using VideoWeb.Common.Configuration;
 using VideoWeb.Common.Security;
 using VideoWeb.Common.Security.HashGen;
+using VideoWeb.Contract.Request;
 using VideoWeb.EventHub.Handlers.Core;
 using VideoWeb.EventHub.Hub;
 using VideoWeb.Services.Bookings;
@@ -32,29 +34,40 @@ namespace VideoWeb
     {
         public static IServiceCollection AddSwagger(this IServiceCollection serviceCollection)
         {
+            var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+            var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+
+            var contractsXmlFile = $"{typeof(AddMediaEventRequest).Assembly.GetName().Name}.xml";
+            var contractsXmlPath = Path.Combine(AppContext.BaseDirectory, contractsXmlFile);
+
             serviceCollection.AddSwaggerGen(c =>
             {
-                var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
-                var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+                c.SwaggerDoc("v1", new OpenApiInfo {Title = "Video Web App API", Version = "v1"});
+                c.AddFluentValidationRules();
                 c.IncludeXmlComments(xmlPath);
-
-                c.SwaggerDoc("v1", new Info {Title = "Video App", Version = "v1"});
+                c.IncludeXmlComments(contractsXmlPath);
                 c.EnableAnnotations();
 
-                c.OperationFilter<AuthResponsesOperationFilter>();
-
-                c.AddSecurityDefinition("Bearer",
-                    new ApiKeyScheme
-                    {
-                        In = "header", Description = "Please enter JWT with Bearer into field", Name = "Authorization",
-                        Type = "apiKey"
+                c.AddSecurityDefinition("Bearer", //Name the security scheme
+                    new OpenApiSecurityScheme{
+                        Description = "JWT Authorization header using the Bearer scheme.",
+                        Type = SecuritySchemeType.Http, //We set the scheme type to http since we're using bearer authentication
+                        Scheme = "bearer" //The name of the HTTP Authorization scheme to be used in the Authorization header. In this case "bearer".
                     });
-                c.AddSecurityRequirement(new Dictionary<string, IEnumerable<string>>
-                {
-                    {"Bearer", Enumerable.Empty<string>()},
+
+                c.AddSecurityRequirement(new OpenApiSecurityRequirement{ 
+                    {
+                        new OpenApiSecurityScheme{
+                            Reference = new OpenApiReference{
+                                Id = "Bearer", //The name of the previously defined security scheme.
+                                Type = ReferenceType.SecurityScheme
+                            }
+                        },new List<string>()
+                    }
                 });
-                c.SchemaFilter<EnumSchemaFilter>();
+                c.OperationFilter<AuthResponsesOperationFilter>();
             });
+            serviceCollection.AddSwaggerGenNewtonsoftSupport();
 
             return serviceCollection;
         }
@@ -78,21 +91,21 @@ namespace VideoWeb
             var container = services.BuildServiceProvider();
             var servicesConfiguration = container.GetService<IOptions<HearingServicesConfiguration>>().Value;
 
-
+            
             services.AddHttpClient<IBookingsApiClient, BookingsApiClient>()
-                .AddHttpMessageHandler(() => container.GetService<BookingsApiTokenHandler>())
+                .AddHttpMessageHandler<BookingsApiTokenHandler>()
                 .AddTypedClient(httpClient => BuildBookingsApiClient(httpClient, servicesConfiguration))
                 .AddPolicyHandler(GetRetryPolicy())
                 .AddPolicyHandler(GetCircuitBreakerPolicy());
-            
+
             services.AddHttpClient<IVideoApiClient, VideoApiClient>()
-                .AddHttpMessageHandler(() => container.GetService<VideoApiTokenHandler>())
+                .AddHttpMessageHandler<VideoApiTokenHandler>()
                 .AddTypedClient(httpClient => BuildVideoApiClient(httpClient, servicesConfiguration));
             
             services.AddHttpClient<IUserApiClient, UserApiClient>()
-                .AddHttpMessageHandler(() => container.GetService<UserApiTokenHandler>())
+                .AddHttpMessageHandler<UserApiTokenHandler>()
                 .AddTypedClient(httpClient => BuildUserApiClient(httpClient, servicesConfiguration));
-            
+
             services.AddScoped<IEventHandlerFactory, EventHandlerFactory>();
             RegisterEventHandlers(services);
                 
@@ -102,8 +115,9 @@ namespace VideoWeb
             };
             
             services.AddSignalR()
-                .AddJsonProtocol(options =>
+                .AddNewtonsoftJsonProtocol(options =>
                 {
+                    options.PayloadSerializerSettings.Formatting = Formatting.None;
                     options.PayloadSerializerSettings.ContractResolver = contractResolver;
                     options.PayloadSerializerSettings.DateTimeZoneHandling = DateTimeZoneHandling.Utc;
                     options.PayloadSerializerSettings.Converters.Add(
@@ -112,7 +126,41 @@ namespace VideoWeb
             
             return services;
         }
-        
+
+        /// <summary>
+        /// Temporary work-around until typed-client bug is restored
+        /// https://github.com/dotnet/aspnetcore/issues/13346#issuecomment-535544207
+        /// </summary>
+        /// <param name="builder"></param>
+        /// <param name="factory"></param>
+        /// <typeparam name="TClient"></typeparam>
+        /// <returns></returns>
+        /// <exception cref="ArgumentNullException"></exception>
+        private static IHttpClientBuilder AddTypedClient<TClient>(this IHttpClientBuilder builder,
+            Func<HttpClient, TClient> factory)
+            where TClient : class
+        {
+            if (builder == null)
+            {
+                throw new ArgumentNullException(nameof(builder));
+            }
+
+            if (factory == null)
+            {
+                throw new ArgumentNullException(nameof(factory));
+            }
+
+            builder.Services.AddTransient(s =>
+            {
+                var httpClientFactory = s.GetRequiredService<IHttpClientFactory>();
+                var httpClient = httpClientFactory.CreateClient(builder.Name);
+
+                return factory(httpClient);
+            });
+
+            return builder;
+        }
+
         private static void RegisterEventHandlers(IServiceCollection serviceCollection)
         {
             var eventHandlers = GetAllTypesOf<IEventHandler>();
@@ -144,11 +192,12 @@ namespace VideoWeb
             };
 
             serviceCollection.AddMvc()
-                .AddJsonOptions(options => {
+                .AddNewtonsoftJson(options =>
+                {
                     options.SerializerSettings.ContractResolver = contractResolver;
-                    options.SerializerSettings.DateTimeZoneHandling = Newtonsoft.Json.DateTimeZoneHandling.Utc;
-                }).AddJsonOptions(options =>
-                    options.SerializerSettings.Converters.Add(new Newtonsoft.Json.Converters.StringEnumConverter()));
+                    options.SerializerSettings.DateTimeZoneHandling = DateTimeZoneHandling.Utc;
+                    options.SerializerSettings.Converters.Add(new StringEnumConverter());
+                });
 
             return serviceCollection;
         }
