@@ -1,5 +1,6 @@
-using Microsoft.AspNetCore;
+using System.IO;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
 namespace VideoWeb
@@ -10,16 +11,23 @@ namespace VideoWeb
         {
             CreateWebHostBuilder(args).Build().Run();
         }
-
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseKestrel(x => x.AddServerHeader = false)
-                .ConfigureLogging((hostingContext, logging) =>
-                {
-                    logging.AddEventSourceLogger();
-                    logging.AddFilter<Microsoft.Extensions.Logging.ApplicationInsights.ApplicationInsightsLoggerProvider>
-                             ("", LogLevel.Trace);
-                })
-                .UseStartup<Startup>();
+        
+        private static IHostBuilder CreateWebHostBuilder(string[] args)
+        {
+            return Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder => {
+                    webBuilder.UseContentRoot(Directory.GetCurrentDirectory());
+                    webBuilder.UseIISIntegration();
+                    webBuilder.UseStartup<Startup>();
+                    webBuilder.ConfigureLogging((hostingContext, logging) =>
+                    {
+                        logging.AddEventSourceLogger();
+                        logging
+                            .AddFilter<Microsoft.Extensions.Logging.ApplicationInsights.
+                                    ApplicationInsightsLoggerProvider>
+                                ("", LogLevel.Trace);
+                    });
+                });
+        }
     }
 }
